Feature/hdrsave

+++ Editor/Resources/DebugColorSpace.shader
@@ -47,10 +47,10 @@
                 return o;
             }
 
-            fixed4 frag (v2f i) : SV_Target
+            float4 frag (v2f i) : SV_Target
             {
                 // sample the texture
-                fixed4 col = tex2D(_MainTex, i.uv);
+                float4 col = tex2D(_MainTex, i.uv);
 
                 #if LINEAR_TO_GAMMMA
                 col.rgb = LinearToGammaSpace(col.rgb);
